+++ src/lib/models/sources/index.ts
@@ -1,4 +1,3 @@
-export { SourceDirectory } from "./directory";
 export { SourceFile } from "./file";
 export { SourceReference } from "./file";
 export { RepositoryType } from "./repository";